Bump version to 1.10.2 and upgrade scala-parser-combinators.
Unfortunately, scala-parser-combinators 1.0.0-RC4 had broken dependency
structure as explained here:
https://github.com/rickynils/scalacheck/pull/64#issuecomment-26704597

There's a new release of parser combinators that fixes the dependency
structure so we switch to it. However, since the 1.10.1 release of
scalacheck depends on broken release of parser combinators we need
a new release of scalacheck with version number bumped.

I've considered various options like releasing scalacheck with
1.10.1-proper or 1.10.1-fixeddeps as a version number but I was given
an advice against using version number qualifiers because various tools
are confused by them. Therefore we have to bump regular number.

This is a big confusing because we are releasing out of 1.10.1 branch
1.10.2 version but I don't see any other option.

The damage is limited because we are still cross-versioning it against
Scala 2.11.0-M6 so most users of scalacheck won't be affected.